Change to show graph per choiceanalysis
